コンポーネントをServerとClientに分ける
