+++ app/pages/top/components/Infos.tsx
@@ -1,0 +1,38 @@
+import InfosClient from './InfosClient';
+
+interface Info {
+    番号: string;
+    投稿日時: string;
+    タイトル: string;
+    内容: string;
+}
+
+const Infos = async () => {
+    try {
+        const res = await fetch(process.env.NEXT_PUBLIC_INFO_API_URL as string, {
+            cache: 'no-store', 
+});
+
+    if (!res.ok) {
+        throw new Error('ネットワークエラーです');
+    }
+
+    const infos: Info[] = await res.json();
+    
+    return (
+    <InfosClient infos={infos} />
+    );
+} catch (error: unknown) {
+    const errorMessage =
+    error instanceof Error ? error.message : '不明なエラー';
+
+    return (
+    <div>
+        <p className="mb-4"><strong>お知らせ</strong></p>
+        <p>エラーが発生しました: {errorMessage}</p>
+    </div>
+    );
+}
+};
+
+export default Infos;
