+++ app/pages/top/page.tsx
@@ -3,13 +3,11 @@ import TextStyle from "@/app/components/text_style";
 import Frame from "@/app/components/frame";
 import Icon from "@/public/icon/44thlogo.png";
 import Image from "next/image";
-import Infos from "@/app/pages/top/components/infos";
 
 export default function TopPage() {
   return (
     <div>
       <h1>Top Page</h1>
-      <Infos/>
     </div>
   );
 }
